+++ tests/profiles/test_profile.py
@@ -34,3 +34,14 @@ class TestUpnpProfileDevice:
 
         # doesn't error
         assert profile._action("RC", "NonExisting") is None
+
+    @pytest.mark.asyncio
+    async def test_icon(self):
+        """Test getting an icon returns the best available."""
+        r = UpnpTestRequester(RESPONSE_MAP)
+        factory = UpnpFactory(r)
+        device = await factory.async_create_device("http://localhost:1234/dmr")
+        event_handler = UpnpEventHandler("http://localhost:11302", r)
+        profile = DmrDevice(device, event_handler=event_handler)
+
+        assert profile.icon == "http://localhost:1234/device_icon_120.png"